// integration_tests/helpers/mock-data/invoice/types.ts
export interface LineItem {
  description: string;
  quantity: number;
  unitPrice: number;
  vat: number;
  subtotal: number;
}

export interface Vendor {
  name: string;
  address: string[];
  vatNumber: string;
}

export interface Customer {
  name: string;
  address: string[];
}

export interface InvoiceData {
  vendor: Vendor;
  customer: Customer;
  date: Date;
  dueDate: Date;
  invoiceNumber: string;
  lineItems: LineItem[];
}